Loan_Bidding_App/requirements.txt
streamlit
pymongo
bcrypt
sendgrid
python-dotenv
pandas
matplotlib
seaborn
